XYZ-59: Implement redis caching layer for Role store.
